Refactor request handling and update query parameters
- Made the `data` field in `HttpRequest` public for easier access in route handlers.
- Renamed `params` to `inner` in `QueryParams` struct and updated related methods for consistency.
- Changed visibility of `insert` and `get` methods in `RequestData` to `pub(crate)` to restrict access to the current crate.
- Updated test module names for clarity by renaming `request_test` to `request` and `response_test` to `response`.
