+++ src/req/request_data.rs
@@ -236,7 +236,7 @@ impl RequestData {
     /// data.insert("key3", vec![1, 2, 3, 4]);
     /// ```
 
-    pub fn insert(&mut self, key: impl AsRef<[u8]>, value: impl AsRef<[u8]>) {
+    pub(crate) fn insert(&mut self, key: impl AsRef<[u8]>, value: impl AsRef<[u8]>) {
         let key = ByteKey::new(key);
         let value = value.as_ref().to_vec();
         self.inner.insert(key, value);
@@ -280,7 +280,7 @@ impl RequestData {
     /// assert_eq!(data.get("missing"), None);
     /// ```
 
-    pub fn get(&self, key: impl AsRef<[u8]>) -> Option<String> {
+    pub(crate) fn get(&self, key: impl AsRef<[u8]>) -> Option<String> {
         let key = ByteKey::new(key);
         self.inner
             .get(&key)
