+++ src/tests/mod.rs
@@ -1,6 +1,6 @@
 mod app_test;
 mod helper_test;
 mod middleware_test;
-mod request_test;
-mod response_test;
+mod request;
+mod response;
 mod router_test;
